+++ src/components/layout/Header.tsx
@@ -1,11 +1,11 @@
-import React from 'react'
-import { Link, useLocation } from 'react-router-dom'
-import styled from 'styled-components'
-import BurgerMenu from './BurgerMenu'
-import Container from './Container'
+import React from "react";
+import { Link, useLocation } from "react-router-dom";
+import styled from "styled-components";
+import BurgerMenu from "../ui/BurgerMenu";
+import Container from "./Container";
 
 const Header: React.FC = () => {
-  const location = useLocation()
+  const location = useLocation();
   return (
     <HeaderDiv>
       <Container>
@@ -18,7 +18,7 @@ const Header: React.FC = () => {
           </LogoWrapper>
           <BurgerMenu />
           <LinkWrapper>
-            {location.pathname !== '/' && (
+            {location.pathname !== "/" && (
               <FavoritesLink to="/">
                 <IconBookmark src="/home.png" alt="home" />
                 <span>Home</span>
@@ -32,15 +32,15 @@ const Header: React.FC = () => {
         </HeaderWrapper>
       </Container>
     </HeaderDiv>
-  )
-}
+  );
+};
 
-export default Header
+export default Header;
 
 const HeaderDiv = styled.header`
   background: linear-gradient(to right, #333, #444);
   position: relative;
-`
+`;
 
 const HeaderWrapper = styled.div`
   padding: 1rem 0;
@@ -48,14 +48,14 @@ const HeaderWrapper = styled.div`
   justify-content: space-between;
   align-items: center;
   color: white;
-`
+`;
 
 const LogoWrapper = styled(Link)`
   display: flex;
   align-items: end;
   gap: 1rem;
   text-decoration: none;
-`
+`;
 
 const IconLogo = styled.div`
   width: 48px;
@@ -63,7 +63,7 @@ const IconLogo = styled.div`
   background-image: url(/logo.png);
   background-repeat: no-repeat;
   background-size: contain;
-`
+`;
 
 const Title = styled.h1`
   font-size: 1rem;
@@ -72,7 +72,7 @@ const Title = styled.h1`
   span {
     color: #f0a500;
   }
-`
+`;
 
 const LinkWrapper = styled.div`
   display: flex;
@@ -84,7 +84,7 @@ const LinkWrapper = styled.div`
   @media (max-width: 650px) {
     display: none;
   }
-`
+`;
 
 const FavoritesLink = styled(Link)`
   display: flex;
@@ -96,10 +96,10 @@ const FavoritesLink = styled(Link)`
   &:hover {
     text-decoration: underline;
   }
-`
+`;
 
 const IconBookmark = styled.img`
   width: 24px;
   height: 24px;
   color: #e0a449;
-`
+`;
