+++ src/pages/Detail.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
-import BookmarkButton from "../components/BookmarkButton";
+import BookmarkButton from "../components/ui/BookmarkButton";
+import Spinner from "../components/ui/Spinner";
 import { Artwork } from "../types/types";
-import Spinner from "../components/Spinner";
 
 const Detail: React.FC = () => {
   const { id } = useParams<{ id: string }>();
